+++ packages/react-searchbox/examples/with-transform-request-prop/src/index.js
@@ -3,4 +3,6 @@ import ReactDOM from 'react-dom';
 
 import App from './App';
 
+import './styles.css';
+
 ReactDOM.render(<App />, document.getElementById('root'));
